Mocked api endpoints are split out
